kt 2019/9

+++ kotlin/src/main/kotlin/io/github/ocirne/aoc/year2019/Day7.kt
@@ -5,9 +5,9 @@ import io.github.ocirne.aoc.permutations
 
 class Day7(val lines: List<String>) : AocChallenge(2019, 7) {
 
-    fun amplifierChain(program: String, phaseSettings: List<Int>): Int {
+    fun amplifierChain(program: String, phaseSettings: List<Long>): Long {
         val amplifiers = phaseSettings.map { IntCodeEmulator2019(program).addInput(it) }
-        var t = 0
+        var t = 0L
         while (true) {
             for (amp in amplifiers) {
                 amp.addInput(t)
@@ -20,17 +20,17 @@ class Day7(val lines: List<String>) : AocChallenge(2019, 7) {
         }
     }
 
-    private fun maximumForPhaseSettings(vararg phaseSettings: Int): Int {
+    private fun maximumForPhaseSettings(vararg phaseSettings: Long): Long {
         return phaseSettings.toList()
             .permutations()
             .maxOf { amplifierChain(lines.first(), it) }
     }
 
-    override fun part1(): Int {
+    override fun part1(): Long {
         return maximumForPhaseSettings(0, 1, 2, 3, 4)
     }
 
-    override fun part2(): Int {
+    override fun part2(): Long {
         return maximumForPhaseSettings(5, 6, 7, 8, 9)
     }
 }
